Add return to fake DbSet add and remove operations
EF returns entities successfully added or removed from a `DbSet` via the
`Add[Range]` and `Remove[Range]` methods. This commit adds the
implementation and unit tests for this behaviour.

Note this behaviour was already implemented in
`FakeItEasyDbSetExtensions.cs` but this lacked test coverage and was not
implemented in the other two fake implementations.